+++ CHANGELOG.md
@@ -10,7 +10,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Remove "addCustomRouter" method. Use "addRouter".
 - Remove "addCustomSetting" method. Use "addSetting".
 - Do not provide the core instance to request handlers. This was made only to maintain temporarily backward compatibility with api plugin, because it is responding with full behaviors collection in deprecated api endpoints, producing a circular reference if the core is saved as a private property of the FixtureHandler Class.
-- Remove "booleanString" option type (--cli=false). Use commander boolean type, which is used appending "--no-" to the option name (--no-cli);
 - Remove "onLoadFiles" method
 - Remove "onLoadMocks" method, use "onChangeMocks"
 - Remove the addition of extra properties when reading files. Define a name for the behavior with a mandatory option.
@@ -35,6 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - feat: Remove deprecated option "feature", "behavior" option should be used instead.
 - feat: Mocks defined using v1 format have to be loaded from folder defined using option "path-v1" instead of "path", which now is used to define the folder from which load routes and mocks in v2 format. Folder defined with "path-v1" option will not be created automatically if it is not found, and the option is not required.
 - feat: Watching files for mocks in v1 format has to be disabled using "watch-v1" option instead of "watch", which now affects only to routes and mocks in v2 format.
+- feat: "watch" option now is a standard commander boolean, so, to disable watch, argument "--no-watch" has to be provided. ("--no-watch-v1" for v1 mocks folder)
+- feat: Remove "booleanString" option type. Now only "number", "boolean" or "string" can be used.
 
 ## [1.6.0] - 2020-12-25